+++ packages/common/src/constants/Config.ts
@@ -9,6 +9,7 @@ export const Config = {
   },
   presentationMode: {
     previousEnabled: 'previousEnabled',
+    viewToggles: 'presentationViewToggles',
   },
   zoom: 'zoomLevel',
   highlight: {